feat: Refactor project to be completely static
- Remove the classes to prevent instantiation, by using top-level
functions instead of static methods
- Update unit tests to use the new way, using module names instead
of using class names
- Update Readme examples for the new usage
