+++ checkarg/none_type.py
@@ -1,10 +1,6 @@
-from .utils import ArgumentNoneException
+from checkarg.exceptions import ArgumentNoneException
 
 
-class NoneType:
-    @staticmethod
-    def is_not_none(value, argument_name: str = None, exception: Exception = None):
-        if value is None:
-            raise ArgumentNoneException(
-                argument_name
-            ) if exception is None else exception
+def is_not_none(value, argument_name: str = None, exception: Exception = None):
+    if value is None:
+        raise ArgumentNoneException(argument_name) if exception is None else exception
